SpectraVis: Visualising and Exploring Spectra Data

This package defines a set of helper function to visualise and explore mass spectrometry data stored as Spectra objects.

SpectraVis functions

plotlySpectra: Interactive visualisation of a single spectrum.

browseSpectra: Browse spectra in a Spectra object.

Examples

f <- msdata::proteomics(pattern = "MS3TMT10", full.names = TRUE)
sp <- Spectra(f)
sp
#> MSn data (Spectra) with 565 spectra in a MsBackendMzR backend:
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           1   4422.62         1
#> 2           2   4422.65         2
#> 3           2   4422.67         3
#> 4           2   4422.74         4
#> 5           2   4422.80         5
#> ...       ...       ...       ...
#> 561         3   4493.77       561
#> 562         3   4493.96       562
#> 563         2   4493.98       563
#> 564         3   4494.14       564
#> 565         3   4494.34       565
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> MS3TMT10_01022016_32917-33481.mzML.gz

if (interactive())
   browseSpectra(sp)

## Use Ctrl+C to interrupt R and stop the application

if (interactive())
   plotlySpectra(sp[1])